pyproject.toml
[project]
name = "ai-detect"
version = "0.1.0"
description = "AI-generated image detection using ensemble of SOTA detectors"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch",
    "transformers",
    "pillow",
    "open-clip-torch",
    "timm",
    "tqdm",
    "rich",
    "typer",
]

[project.scripts]
ai-detect = "ai_detect.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_detect"]
